requirements.txt
boto3==1.24.0
botocore==1.27.0
python-dotenv==0.20.0
pygelf